requirements.txt
django
django-rest-framework
django-ipware
protobuf
protobuf3-to-dict
